+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add schedule confirmation and explanation message
 - Add ingredient info in formula making screen
 - Add option to create a new formula category from the formulas list page
+- Add low stock in formula making screen
 
 ### Version 5.9
 - Introducing formulas import/export in JSON format