+++ releasenotes.md
@@ -4,3 +4,4 @@ Whats New:
 - Add schedule confirmation and explanation message
 - Add ingredient info in formula making screen
 - Add option to create a new formula category from the formulas list page
+- Add low stock in formula making screen